Moving Navigator ScreenModel to ScreenModel module
