+++ voyager-screenmodel/build.gradle.kts
@@ -16,6 +16,7 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 api(projects.voyagerCore)
+                api(projects.voyagerNavigator)
                 compileOnly(compose.runtime)
                 compileOnly(compose.runtimeSaveable)
                 implementation(libs.coroutines.core)
